feat: introduce plugin-based notifier system and refactor core
