+++ VersaLog/plugin/notify.ts
@@ -1,0 +1,11 @@
+import notifier from "node-notifier";
+
+export const notifyPlugin = (log: any) => {
+  log.setNotifier((title: string, message: string) => {
+    notifier.notify({
+      title,
+      message,
+      appName: "VersaLog",
+    });
+  });
+};
