+++ package.json
@@ -1,34 +1,19 @@
 {
   "name": "versalog",
-  "version": "2.0.0",
+  "version": "2.1.0",
   "description": "A simple and customizable logger for Node.js",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "main": "dist/core/index.js",
+  "types": "dist/core/index.d.ts",
   "author": "kayu0514",
   "license": "MIT",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "lint": "eslint .",
-    "format": "prettier --write .",
     "build": "tsc"
   },
-  "keywords": [],
-  "devDependencies": {
-    "@eslint/js": "^9.36.0",
-    "@types/node": "^24.0.10",
-    "@types/node-notifier": "^8.0.5",
-    "eslint": "^9.36.0",
-    "globals": "^16.4.0",
-    "jiti": "^2.6.0",
-    "prettier": "^3.6.2",
-    "typescript": "^5.8.3",
-    "typescript-eslint": "^8.45.0"
-  },
   "dependencies": {
-    "date-fns": "^4.1.0",
-    "luxon": "^3.6.1",
-    "node-notifier": "^10.0.1",
-    "pretty-format": "^30.0.2",
-    "versalog": "^1.0.0"
+    "date-fns": "^4.1.0"
+  },
+  "devDependencies": {
+    "@types/node": "^24.0.10",
+    "typescript": "^5.8.3"
   }
 }
