+++ tsconfig.json
@@ -1,5 +1,6 @@
 {
   "compilerOptions": {
+    "rootDir": "./VersaLog",
     "outDir": "./dist",
     "declaration": true,
     "module": "commonjs",
